requirements.txt
# Galatiq Invoice Pipeline — Dependencies
# In production, we'd use Poetry for dependency management and lockfiles.
# Install: pip install -r requirements.txt

# API
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
httpx==0.27.0

# Validation & config
pydantic==2.8.2
pydantic-settings==2.13.1

# LangChain / LangGraph
langchain-core==1.2.17
langchain-openai==1.1.10
langgraph==1.0.10
langsmith==0.7.12

# LLM providers
anthropic==0.84.0
openai==2.26.0

# Document parsing
pdfplumber==0.11.9
PyMuPDF==1.27.1
pytesseract==0.3.13
Pillow==10.4.0

# Data
pycountry==26.2.16

# Testing
pytest==7.4.4